spatial: Fix behavior after USB reattachment
The Phidged Spatial never recovered after detaching and reattaching to
the USB port. This commit fixes that.

The cause of the failure was the following:

* After reattachment, the device time stamp restarts from 0. This caused
our driver to throw the following error messages:

        [ WARN]: Time went backwards [...]! Not publishing message.

* However, the data interval is also reset to the default of 256 ms. If
the parameter data_interval_ms is set to something else, this caused the
arriving data to always be outside the acceptable window for
synchronization. Therefore, synchronization never happened, and the
driver never resumed publishing messages.

This commit fixes the bug by setting the appropriate data interval after
each reattachment and forcing a resynchronization immediately.